[KGEN] Register Passable SLValues use RValueType for Field and Init Type (#20057)
This patch updates the default capture behavior of SLValues whose
RValueType is register passable. Consider this case:

```
fn foo(owned y:Int):
  var w = 5
  fn bar() escaping -> Int:
     y = y + 1
     w = w + 1
     return w
```
IR:
```
	lit.func @"foo($builtin::$int::Int)"(%y: !kgen.declref<@Int>) -> !lit.none {
  	%y_0 = lit.varlet.decl "y", var = true, synth = true : <@Int>
  	pop.store %y, %y_0 : !pop.pointer<@Int>
  	%w = lit.varlet.decl "w", var = true, synth = false : <@Int>
  	%0 = kgen.param.constant: @Int = <#lit.struct<{value = 5}>>
  	pop.store %0, %w : !pop.pointer<@Int>

  	lit.func *"bar()"() escaping -> !kgen.declref<@Int> attributes {specialFnKind = 0 : i8} {
    	%3 = pop.load %y_0 : !pop.pointer<@Int>
    	%4 = kgen.param.constant: @Int = <#lit.struct<{value = 1}>>
    	%5 = kgen.call @Int::@"__add__"(%3, %4) : (!kgen.declref<@Int> borrow, !kgen.declref<@Int> borrow) -> !kgen.declref<@Int>

    	pop.store %5, %y_0 : !pop.pointer<@Int>
    	%6 = pop.load %w : !pop.pointer<@Int>
    	%7 = kgen.param.constant: @Int = <#lit.struct<{value = 1}>>
    	%8 = kgen.call @Int::@"__add__"(%6, %7) : (!kgen.declref<@Int> borrow, !kgen.declref<@Int> borrow) -> !kgen.declref<@Int>
    	pop.store %8, %w : !pop.pointer<@Int>
    	%9 = pop.load %w : !pop.pointer<@Int>
    	lit.return %9 : !kgen.declref<@Int>
    	lit.end_func
  	}
```
Old behavior:
```
lit.struct.decl @"_CI_” {
  lit.struct.field field0 : !pop.pointer<@"$builtin"::@"$int"::@Int>
  lit.struct.field field1 : !pop.pointer<@"$builtin"::@"$int"::@Int>
  lit.func @"__init__”(%self: !pop.pointer<@CI>, %field0: !pop.pointer<@Int>, %field1: !pop.pointer<@Int>) {
    // copies here
  }
}

```
Note that the captured slvalue y_0 results in a pointer field type whose
value is copied from the initializer argument (effectively a copy by
reference).

Note that the captured slvalue w also results in a pointer field type
whose value is copied from the initializer argument (again, effectively
a copy by reference).

In the first case, %y_0 refers to %y which is owned by the function.
Capturing a value owned by the function should result in a copy since
the closure’s lifetime is equal to or greater than the function AND its
default capture by copy, not capture by reference.

In the second case, %w is declared as a local mutable reference which
should translate to a member mutable reference in the class.

Should all slvalues result in a RValueType field + init with the
expectation that the original values are loaded before initialization?
I think the answer is yes. For embedded mlir ops this can still result
in unsafe captures but I believe the behavior is correct. i.e. consider
```
fn make_pointer() -> __mlir_type.`!pop.pointer<index>`:
  let alignment = 0
  let size = 8
  return __mlir_op.`pop.aligned_alloc`[
          _type : __mlir_type.`!pop.pointer<index>`
      ](alignment.value, size.value)

fn foo():
  var q = make_pointer()
  fn bar() escaping -> NoneType:
     __mlir_op.`pop.aligned_free`(q)

```
In this case “q” is an SLValue with RValueType of pointer<index> and
type pointer<pointer<index>>. The expectation is that the field type in
the closure is pointer<index> and its value is copied from the
pointer<index> init arg. This will result in unsafe capture but we allow
that since we allow the use to use !pop.pointer directly.

MODULAR_ORIG_COMMIT_REV_ID: d1c4bb926c92d86052d9cd627c949174313eb46f