[MOGG] Allow sideloading user provided kernels (#20095)
Adds a flag to the graph compiler which allows for user provided kernels
to be sideloaded during compilation. These are exactly the same as the
existing kernel library code which is compiled in so they can replace mo
ops as usual. The user kernel library is renamed to avoid collisions and
then appended to the main library for compilation.

MOGG_ORIG_REV_ID: 73f8eae0f157443dca68bba45fa903e6dc97f048