[Stdlib][Profiling] Wait on existing out_chain in sync_parallelize (#20301)
There is no need to create a local chain in sync_parallelize. Creating a
new chain does not fix the race condition mentioned in the comment. When
mark_ready is called on the out_chain, we are still in a mojo call stack
where the runtime may be needed.

By waiting on the existing chain, traces for the sub-tasks are kept,
which fixes [Internal link]

ALGORITHM_ORIG_REV_ID: 69edae7c1ce92b522c2b22a79ed2600b6994ea06