[mojo-vscode] Add initial debugger integration (#20200)
This PR integrates lldb-vscode as a debug adaptor for mojo programs.
This will make it easier to provide a good debugging experience for
mojo within vscode. The starting point in this PR is to add another run
command that allows for starting a debug session for the current mojo
file. This session runs via the jit (`mojo run`), which will allow for debugging
the mojo program *and* the mojo executable at the same
time (should be transparent to users, but nice for developers that want
to quickly debug the mojo executable itself).

The changes to the package.json are basically a copy of what the private
`lldb-vscode` extension defined in lldb provides.

MODULAR_ORIG_COMMIT_REV_ID: dbfd405d575d1a8cfbac7f16fd05fcb2bfc0e321