[REPL] Autoindent the source code (#20226)
Fortunately, Mojo has very simple scoping rules, so having a
heuristics-based autoindent feature is not very difficult. The rules of
these heuristics are described as a comment in the source code.

One caveat is that LLDB has a setting for the tab size, which is
currently 4 and we use it to set the indent level. I'm going to propose
upstream to use 2 as default, as 4 is too much. The reason is that the
REPL transforms tabs into spaces, so whenever you want to deindent, you
need to delete at least 4 characters, which is a bit too much.

MODULAR_ORIG_COMMIT_REV_ID: f8586378609d87c5c0d1d7215720a85b9c4f0f7c