[MojoLLDB] Remove lldbinit file (#19507)
It conflicts with the `plugin load` command issued by `mojo repl`. It's
not clear to me if this was ever supposed to work and it was working
just by chance or not, but Walter was out so I was unable to get an
answer. Aman thinks the `lldbinit` file is obsolescent, and removing it
should fix the issue as well, so this PR removes the `lldbinit` file,
and further, causes CMake to delete it if it is there so an explicit
`clean` is not necessary.

MODULAR_ORIG_COMMIT_REV_ID: 402f256d315d6eb862f9c05f422ae0cfdf5ba427