[Support] Fix target info test for x86_64 macOS (#20406)
The test assumes that all hosts for which `__APPLE__` is defined are
Apple Silicon, but in fact they may be x86_64.

This is a part of fixing the `buildAndTestMacOS` GitHub workflow, but
other test failures remain and will be fixed in future commits.

Related to issue #20404.

MODULAR_ORIG_COMMIT_REV_ID: 79a61dadf6ef46b78ca23e27408b75a8529ea7e5