[modular-tool] Add CAS IDs to the InstalledPackageManifest. (#20310)
This patch adds CAS IDs to the files written to disk stored in the InstalledPackageManifest. This will be important later for the update path, so we don't re-download files that have not changed.

This patch also adds a version to the InstalledPackageManifest so that if we need to change it in the future we have a way to continue to parse old versions of the manifest.

MODULAR_ORIG_COMMIT_REV_ID: cf99268076689efbdc46da97a51d6b88461157b1