[Mojo] Refactor Capture to Include AnyValue (#20510)
This diff makes it so that the Capture instance instead stores an
AnyValue and recovers the mlirValue from the AnyValue by storing the
value type. The idea is that its cheaper to remember your type then to
dynamically cast N times for N potential value types and that it's
better to uncover the mlirValue when needed than to store the values
twice for the lifetime of Shared.

Context:
To implement closures, we need an AnyValue (to emit operands from
ExprEmitter) and an mlir Value (to replace occurrences of capture with
extracted member from closure struct) from a Capture instance. Currently
we store the Mlir value in Capture. This requires us to search the value
type by casting the ASTDecl in order to create an AnyValue when we emit
the initializer call.

MODULAR_ORIG_COMMIT_REV_ID: 2c486513d6653b2596c1146f1962669ec09ab28a