[KGEN] Move `Support/Interpreter` into the KGEN directory (#20594)
The interpreter has grown into having a large number of Mojo-specific
implementation detail. It is no longer suitable for general use. To
reflect this evolution and solidify the ownership of the component,
move the interpreter into the KGEN directory.

As part of this PR, I added an `InitAllKGENDialects` target, because
this required adding yet another dialect extension (for the `DebugInfo`
dialect). This also pulls a bunch of logic out of the header file.

Part of addressing [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 787be72b612f8fc729b235bb660159345cb4798e