[Support] Some HostMachineInfo/TargetInfoAttr/TargetInfo/PackageTarget rationalization (#20635)
This is a cleanup to exploit the new TargetInfo POD struct and remove
some (but not all) code dup.

- We take getMArchFeatures in MArchTarget as the definitive way to
construct the TargetInfoAttr from either a given arch/cpu/tune or for
the host. This is implemented without deps on HostMachineInfo. The
library continues to depend on clang.

- But we split that so it's possible to stop at the POD TargetInfo using
getMArchTargetInfo (and getHostTargetInfo for the host).

- We also move a few other host-related helpers in there, such as
getTargetMachineForHost and getHostCPUFeatures.

- Much of the FilterTarget.cpp logic is now in TargetInfo, so redirect
there (I was careful to copy it earlier).

 - A bunch of serialize/deserialize round trips removed.

- Remove JSON and host-related stuff from MAttr.cpp, which is now just
about supporting the attr.

There's still cleanup to be done for HostMachineInfo, in particular to
remove its 'cpuFeatures' since that's the big remaining code dup issue.
And we should move getMArchFeatures into a new MArchTargetWithMLIR
library since it's the only MLIR dep.

But this is at least a step forward.

MODULAR_ORIG_COMMIT_REV_ID: 3784eef2092339a855395f9fefee62222b6ed549