[KGEN] Add a terminator op for `pop.coroutine.await` (#20657)
As the only region operation in the IR with no terminator, this op has
caused some level of inconvenience throughout the stack. I remember the
original reason why `pop.coroutine.await` did not get a terminator was
due to the parser hack introduced for embedding MLIR regions in MLIR ops
that predates the `__mlir_region` keyword: the terminator operation
would conflict with the generated `lit.end_func` and cause a verifier
error in the parser prior to `lower-semantic-cf`. That hack is long
gone, so it is no longer an issue to give this op a terminator.

The rest of the FIXMEs will be addressed in a follow-up.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 1b1781c663420ddf2723b2fe9d9181170071b470