[Support] Fixup unknown target triple vendors (#20688)
When invoking `mojo package -march armv8.2-a -mcpu neoverse-n1` on a
2021 Apple MacBook Pro with an Apple Silicon M1 chip, I noticed that the
built package's module's target triple was set to "aarch64-a-macosx". In
this triple, the "vendor" is set to "a", which is not a known vendor.

An unknown vendor in a triple isn't in and of itself an error, but in
this case it's clearly a bug: the hyphen in `-march armv8.2-a` is
confusing LLVM's triple parsing logic as we build the triple in
`M::getMArchTargetInfo`, resulting in the "-a" being treated as the
vendor component of the triple.

To prevent triples like this from being produced, re-set the vendor
component of the triple toward the end of the building process. Instead
of simply checking that the vendor string is empty, check that it's
unknown (which covers the empty case as well), and re-set it if so. This
updates the underlying string in the triple, from "a" to "unknown".

This patch also updates a use of "armv8.2-a-macosx" in the KGEN tests
(not a valid triple, as "armv8.2" isn't a canonical architecture in LLVM
triples, and "a" is not a known vendor).

MODULAR_ORIG_COMMIT_REV_ID: 2bfcce40e3ce9214dd3bd8a58280b7463cd8e30c