[MO] Adding Padding as an argument to Stencil Ops (#20696)
This adds `padding` as an argument to mo.conv, mo.avg_pool, and
mo.max_pool, and updates the shape functions, lowerings, and frameworks
to properly support it, along with a bunch of tests. Sorry for all of
the tests.

It includes tests for the shape functions, but does not yet Lower
framework padding to mo, frameworks are simply making the constant zeros
explicit via this patch. (#20790, #20792)

Those gaps will be address by a future PRs.

MOGG_ORIG_REV_ID: 7ae7e6d8562bd912645758b6a5737ca284dc0395