[Support] Add `ASN.1` module with `ObjectID`. (#20672)
This patch adds an ASN.1 module for the modular support library, and populates it with an ObjectID structure. This will be necessary for entitlements, because we plan to use X.509 extensions which are identified with ASN.1 Object Identifiers.

The ObjectID class contains some small optimizations that apply to Modular-namespaced arcs - we will usually be able to avoid encoding over half the numbers in the OID each time. The class also supports non-Modular-namespaced OIDs, we just have to store every node.

MODULAR_ORIG_COMMIT_REV_ID: defedb0682b5fadb2365f6996b08046e1f1c4133