[Support] Add `Random` module. (#20854)
This patch adds a `Random` module to Support that can be used for generating cryptographically-secure random bytes. This is important to own because LLVM doesn't provide a CSPRNG that is documented as such, and so the implementation could change at any time.

MODULAR_ORIG_COMMIT_REV_ID: f339cfee4c0b4041335732ca7882558f504ba38d