[mojo] Only set target info on elaborated module (#20973)
When building a Mojo package, an MLIR module is created to represent the
package contents, and is serialized as both pre-elaborated and
elaborated bytecode.

The pre-elaborated bytecode is meant to be target-agnostic: when a
package is pre-built for a specific target, and a user of that package
is building for a different target, the pre-elaborated bytecode is
loaded and then elaborated for the user's target.

However, prior to this patch, the pre-elaborated module bytecode would
contain target and build information. As a result, attempting to
elaborate this module for a different target would trigger an elaborator
error: "target did not match."

As a result, this commit modifies `mojo package` to only set the target
and info immediately before elaboration.

Addresses a portion of #19983.

MODULAR_ORIG_COMMIT_REV_ID: de048a8e853b4a084f040a566d95b3442afc8904