[KGEN] Don't export package's "inflated" ops, NFC (#20987)
When building a Mojo package, an MLIR module is created to represent the
package contents, and is serialized as both pre-elaborated
(target-agnostic) and elaborated (for a specific target) bytecode.

When the target that the elaborated bytecode was built for and the
target of the module importing the package don't match, then the
pre-elaborated bytecode is used. Currently, pre-elaborated bytecode is
"inflated" into the module being compiled, meaning the package's ops are
simply inserted into the module being compiled, as if they were part of
that module. They're then elaborated along with the rest of the module,
and things mostly "work."

One quirk here is that the "inflated" ops are all marked as not being
exported -- they may have been exported from the original package, but
they're not being exported from the module they're being "inflated"
into.

However, in a future commit, this "inflation" will be augmented with an
option to actually compile the package on-demand, and so whether an op
was exported or not needs to be preserved in the pre-elaborated
bytecode. This commit preserves the export kind on the pre-elaborated
bytecode, and instead sets ops to not-exported during inflation.

Addresses a portion of #19983.

MODULAR_ORIG_COMMIT_REV_ID: ae759f4fa529fcea0f036ca511c047bd85690379