[mojo-lang] Start staging in argument lifetimes, NFC. (#21049)
This does a small amount of exploratory work to stage in support
for having memory argument get implicit lifetimes.  The most
significant piece of this is to stage in a
ASTType::getReferenceElementType
helper (used when dealing with arguments that should be references)
relaxing some KGEN verification to tolerate these, and adding some
(disabled!) code to enable this that we can play with.

This is still a far way away from this being useful.  We'll need
more aggressive staging mechanisms, and argument conventions are
not the right place to start unless you want to shake the tree
hard.

MODULAR_ORIG_COMMIT_REV_ID: 2bf19614f2fb8aa01d76533f11342cbec789be39