[Mojo] Replace the Constant Symbol Attr With a Custom Attribute (#20729)
The verifier of a symbol verifies that the contents of the symbol
attribute refer to a function. The ClosureEmitter was abusing this
attribute because the closure signature represents a function interface,
it does not refer to a function.

MODULAR_ORIG_COMMIT_REV_ID: cf997a142926235bf88dbcb53639ff2e79ffa493