[Mojo] Simplify `InputParamBindings::verifyBindings` (#21061)
The existing signature made it seem like this method depends on the call operands, but it doesn't. Also, before there was one `verifyBindings` function that accepted all fields necessary for making and emitting an error message in the case that the provided parameter values cannot be bound to the target input parameters. This PR changes it so there are two `verifyBindings` functions: the first contains the actual check logic and the second is a wrapper around the first that defines callbacks that emit diagnostic messages upon failure. The resulting error information is now returned in the fitness of the proposed binding instead of via reference arguments.

MODULAR_ORIG_COMMIT_REV_ID: 3e503e5d53f1eae6c38a402942984018afc8292f