[mojo-repl] Refactor diagnostic emission to reference the input expression (#21023)
Due to the lack of top-level code in mojo, the REPL currently does some magic
to mimic support by wrapping the input expression in a function. This works
kind of okay, but one big problem is that location information coming from the
parser references the wrapped expression, not the input expression. This makes
diagnostics confusing, breaks any kind of language tooling, etc. This PR starts
to mitigate that by building a mapping from the wrapped expression to the input
expression, and using that to remap locations from diagnostics. As part of this,
we can enable cleaning up a nice chunk of the current hacky logic, even if we
can't yet remove it all.

Whenever mojo supports top-level code, we should be able to rip all of this out
and just use things naturally from the parser.

MODULAR_ORIG_COMMIT_REV_ID: 7e78ed5d4f2873b77ee1b6a0a72cfc2c6a8fa879