[KGEN] De-dupe createResourceAttr in Support, NFC (#21173)
Both mojo-package.cpp and kgen-opt.cpp create an attribute for the
MLIR bytecode of a module (the latter does so for testing purposes).
On-demand package compilation will also soon be modified to create such
an attribute.

Extract the logic out to a Support routine now, to remove the duplicated
codde and simplify the upcoming patch.

MODULAR_ORIG_COMMIT_REV_ID: 6065ab232862310ce4d008636cee589fcbdbffea