[Telemetry] Remove telemetry file writer assumption from unit tests (#21184)
This changes the telemetry unit tests so that they now look for the
presence (or lack thereof) of a specific message in the telemetry file.
Which also removes the assumption that the test is the only process
writing to that file.

This is to address a flake in CI
[Internal link] where it looks
like a mojo process can run at the same time as the telemetry unit tests
in CI.

MODULAR_ORIG_COMMIT_REV_ID: 393338c8412cfa2fb75ce2ecc77c7d5b4f8345e2