[KGEN] Un-indent LowerPreElaboratedLITPass, NFC (#21200)
Split the declaration of `LowerPreElaboratedLITPass` and its
implementation of `runOnOperation`. This allows the body of that (long)
function to be un-indented one level, making it (arguably) easier to
read.

This simplifies the changeset in an upcoming patch that implements
on-demand compilation of packages, which is described in #19983.

MODULAR_ORIG_COMMIT_REV_ID: 98c89023a1774e034696088636e7f61d4867e4fd