[ModelServing] Add OTEL logs and traces to Mutt (#21160)
This specifically does not use the existing TelemetryContext type in use
in other places in the codebase. That type bakes the transport of
observability too tightly the collection & would require a large
refactor for mutt to use it. It may make sense to do that refactor, but
that work isn't necessary yet.

- Add instrumentation for logs & traces
- Add traces to all endpoints

MODULAR_ORIG_COMMIT_REV_ID: baf2a99ffa0c04705dbbd34adc366a689826cbce