[KGEN] Null Check Impl In Dtor (#20847)
We do not want to load the impl to call its destructor if it is null. It
will be possible for the impl to be null in an upcoming PR when I null
out the impl of the existing in Closure Wrapper's moveinit.

MODULAR_ORIG_COMMIT_REV_ID: 201caac53138513d2e13c8d62e9bb90143332dd3