[LLDB] Install missing linux deps (#21393)
On Linux we've been building LLDB without libedit, which causes it not
to have a nice experience in the command line, as reported in
[Internal link]

A simple fix is to follow the LLDB build guide

https://github.com/llvm/llvm-project/blob/main/lldb/docs/resources/build.rst?plain=1#L73
which mentions libedit-dev and libncurses5-dev, so I'm adding them to
the list of install dependencies.

We need to be careful to include the corresponding deps when shipping
for other distros.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 4ed91f89b3eed727e74337ddad095d94971bb3ee