[Mojo] Add `escaped_identifier` token kind and fix diagnostic printing of escaped identifiers (#21390)
The source range of escaped identifiers would cause a crash when being emitted, because the information about whether the identifier was escaped or not was not propagated correctly. This patch fixes that by differentiating between regular and escaped token kinds, and propagating it through the appropriate expression nodes.

Huge kudos to @Mogball for helping to diagnose and minimally repro the problem!

MODULAR_ORIG_COMMIT_REV_ID: 946293a4d08b1f75878d4baacece687ae9710a6b