[Mojo] Add @nonmaterializable decorator (#21346)
The @nonmaterializable decorator is for use with literal types (or at
least parameter-domain types) that can auto-convert to a runtime type,
but which have different semantics in the parameter domain. The
motivating example, which is not in this commit, is infinite precision
integer literals that, unless conversion to another type is performed,
automatically convert to the Int runtime type when materialized.

MODULAR_ORIG_COMMIT_REV_ID: 80633c1842bc0736f7c1a6229b0a8e8c7588729f