[mojo-lang] Introduce a new `lit.varlet.decl2` op. (#21477)
This introduces a new operator for defining local stack storage,
which is just like lit.varlet.decl, but which defines a lifetime
and returns a mutable `!lit.ref` type.

This is basically a different approach to the patch in #21029,
which makes it return a `!lit.ref` type to keep the pointer and
lifetime together all the time. This will eventually be useful
to CheckLifetimes and the parser.

MODULAR_ORIG_COMMIT_REV_ID: 9a595ff4fa5a225e66ec776c315145b3007a1246