[Mojo] enable IntLiteral in the parser (#21448)
### [Mojo] enable IntLiteral in the parser

This enables IntLiteral as the type of literal integers in the parser.

IntLiteral is an infinite precision nonmaterializable type.  So
intermediate computations that happen at compile time involving
integers can now exceed the 64-bit bound of the Int type.  A famous
example is `-(2<<62)`, which has an end result that equals the minimum
signed 64-bit integer, but which produces an intermediate value that
is out of bounds (maxInt + 1).

This commit also adds tests for IntLiteral, as well as changing many
tests where literals were used.  The changes were typically either:
update a check to show that the type is now IntLiteral, or explicitly
cast the IntLiteral to Int, such as in the case where the index value
of an Int is desired.

MODULAR_ORIG_COMMIT_REV_ID: ed5cdae42f6b0240dd212196e7582350865ef960