[Support] Add symmetric cryptography. (#21356)
This patch adds basic symmetric cryptography primitives. The API is designed to be difficult to mis-use and accidentally compromise the secrecy of the message being encrypted.

Symmetric keys *can* be written out to DER form (and encoded to PEM if desired), but users should keep in mind that it is key material and so must be kept secret!

MODULAR_ORIG_COMMIT_REV_ID: 838bf1283ced393d7b7bae5c7adf9acf2d2578d4