[mojo-lang] Make `lit.varlet.decl2` loadbearing and fix issues. (#21486)
This switches one part of the Mojo parser (makeArgLValueVarSlot)
over to unconditionally using the lifetime-enabled vardecl support
and fixes issues that come up as a result of that.

These issues include adding support for:
 - Builders and other helper methods for VarLetDecl2Op.
 - LifetimeTrackable support for VarLetDecl2Op/RefToPointer.
 - CheckLifetimes support for all the new stuff.
 - Switches `makeArgLValueVarSlot` to use this.
 - Updates a bunch of testcases.

With this in, we should probably be able to move a bunch of other
stuff over, but probably want to add support for references (e.g.
GEP on a reference, load/store of a reference etc) before we do
that.

MODULAR_ORIG_COMMIT_REV_ID: 4f40d9dc8e22849d21104311d253615b4b6f6818