[KGEN] Fix SROA incorrectly decomposing stack allocations (#21582)
SROA is decomposing "stacks" in the form of `%0 = stack_allocation N x
T`
where `N > 1`. It does so even if there are uses of `%0` that are
`pop.offset` with constant indices, but it does not check the users of
the offset. This is incorrect, because `pop.offset` returns an alias of
the original pointer and is not an element access -- its users need to
be checked to ensure the pointer does not escape.

Fixes #21576

MODULAR_ORIG_COMMIT_REV_ID: 2d4eeb5c1f508060d911dfd43dfafe38901fcd3d