[mojo-lang] Move nested function lowering back to `lower-lit` (#21573)
Lowering of nested functions has been moved back to lower-lit from
lower-semantic-cf. There was some reason why it was originally
moved up to lower-semantic-cf, but whatever it was, it doesn't seem
to be relevant anymore. This is necessary because the
LITSignatureType has to be preserved for nested functions in
check-lifetimes, which ParamDeclareRegionOp will lack after the split.

MODULAR_ORIG_COMMIT_REV_ID: f81ee07f51fe24e2cced0ee3196c75569699aa64