[KGEN] Fix ModuleOp leak in KGENPackage unit test (#21596)
When the KGENPackage unit tests are run with sanitizers, such as in
[Internal link]
a leak of the `ModuleOp` created in the tests is reported. This is
because the module is never freed; an `OwningOpRef` can handle freeing
the op at the end of its scope.

MODULAR_ORIG_COMMIT_REV_ID: b57ad80ddf7e31ad5ac5a81a40c854889ef9cbcb