[LLDB][mojo] Add a decorator for visualizing simple wrapping types. (#21453)
Many types in mojo are simple wrappers of the real underlying value the
user is interested in when debugging, e.g. Int, StringLiteral, Pointer,
etc. As a way to simplify the visualization of these variables, I'm
adding a utility decorator `@lldb_formatter_wrapping_type` that, whenver
applied to a struct, indicates LLDB to create a display the first field
of the variable instead.

As these types tend to be the majority of types (say Int), we need to
use a static approach like this one to ensure performance of the
debugger.

In this patch I'm starting with Int and StringLiteral, and once approved
I'll go ahead with others.

This a sample old output
<img width="441" alt="Screenshot 2023-09-15 at 11 20 33 PM"
[Internal link]

And this is the one with the decorator

<img width="369" alt="Screenshot 2023-09-16 at 12 29 12 AM"
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: d19294c02a200e394ac946ce1379c595eb69b7c8