[modular-tool] Replace `-installer-outfile` flag. (#21799)
This patch removes the -installer-outfile implicit behavior and replaces it with the explicit behavior of using a static replacement for the string `$_self.cli_outfile`. This enables us to more cleanly delineate in scripts when an output file is required, and it also cleans up some tricky behavior that was causing bugs by looking for `python` in the executable string.

MODULAR_ORIG_COMMIT_REV_ID: a13b8f8015aa44aea44685fc29cd7ee4471a391b