[KGEN] Use `ExportMap` typedef (NFC) (#21847)
The `ExportMap` typedef was introduced in #10686, and while it was used
in the declaration of the `getExportedSymbols` function, it was not used
for its definition. Use it for the definition as well, so any futurte
changes to the type are propagated there as well.

MODULAR_ORIG_COMMIT_REV_ID: b1b6a119cdd5a404e96f56dd50e9e0184061a517