[CUDA] Another go at CUDA-specific options handling (#21900)
This is all so that the MO AddDevicesPass can interrogate the host
machine for available devices, including CUDA devices, and attach that
to the mo.graph to guide lowering. With this change, AddDevicesPass uses
exactly the same code as we use at runtime to check that the model's
device requirements can be satisfied by the host machine, again
including CUDA devices.

- Rip out the 'global context objects' from Runtime, it was wrong
headed.
- Rejigg mof.cpp to clearly gather the options, then install them into
the runtime.
- Plumb an (optional) runtime into AddDevicesPass. If available, use it
to gather a full description of the host.
- I keep the --enable-cuda cl option to control whether AddDevicesPass
sees a runtime. But it is no longer needed elsewhere, and AddDevicesPass
now knows nothing about CUDA.

Follow ups:
- As already recognized, the mix of cl vs mlir pass vs programmatic
options is an unholy mess.
- We need to clearly distinguish command-line / API options which apply
to all models in all frameworks from those which should only effect the
current compilation.
- We don't have a way to convey per-model execution options, and it's a
big refactor to plumb them all the way down into the MEF executor (I
tried!). So suggest we steer clear of that unless a very clear use case.
- We have a mix of `std::unique_ptr<llcl::MemoryBuffer>` and
`llvm::MemoryBufferRef` in the driver which should all be replaced with
`M::BufferRef`.
- The driver insists on creating a runtime per model group to be
executed, but I think we should be creating a global runtime and share
it between running frameworks, both for compile actions and running
models. In other words, the XXContext object in the driver is the one to
hold the runtime, and there's driver level protocol for establishing it.
That's a CXXDriverAPI change.
- I create a runtime 'by hand' for running the MO compilation passes,
but it is only used by AddDevicesPass. There are many passes which need
a runtime, and create one if none is given. We should plumb the runtime
down into those too.

MODULAR_ORIG_COMMIT_REV_ID: c454b7116eb0e0153da4e6ab07076c9e1a6dfc06