[LLDB] Move expression logger out as its own class. (#21832)
For context, we've built an expression logger inside the MojoTypeSystem.
This logger is tied to an specific target, just like the type system for
expressions we normally use in the REPL. This has been an okay-ish hack
so far, but in order to be able to create TypeSystems for bare modules
(i.e. no target available, therefore no logger available), which is used
in normal debugging, we need to get move the logging functionality out
of the type system.

This has some benefits in terms of clean code, e.g. the Jupyter Kernel
doesn't need to know that there's a type system anymore.

MODULAR_ORIG_COMMIT_REV_ID: 5b6bc18c845b3eec693a949e0d89649796254d53