[KGEN] Make sure to load the NVVM dialect when generating attributes (#21935)
"Dialect attributes" in MLIR are weird. Creating a dialect attribute
does not load the associated dialect, but parsing one does. In
addition, the dialect has to be loaded for its interfaces to be as well.
Ensure that the NVVM dialect is loaded as a dependency dialect of
LowerKGENToLLVM so that its translation interface kicks in the LLVMIR
export.

MODULAR_ORIG_COMMIT_REV_ID: 92b7e77bea48a8e6a6981f8f404a518f502c7f36