[KGEN] Use memset to compress how memory blobs are materialized (#21964)
This prevents compile time death when a large pointer is allocated and
returned in an parameter expression by "compressing" large sequences of
stores with `llvm.intr.memset`.

There are many ways to further improve this: just allocating a pointer
at runtime doesn't result in a memset to 0 of the whole blob. In order
to have the same behaviour in the parameter domain, the interpreter and
memory space attribute will need to keep a transaction history of all
the
stores and treat the rest of the memory as "garbage", but that's for
another day.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: dd73e2ab17d7039477700f5f151dba6c368aad4f