[KGEN] Add `LIT::PackageArchiveAttr` (NFC) (#21938)
Add an attribute that represents an elaborated module and its standalone
archive, and make use of that attribute in `LIT::PackageLinkOp`. A
future commit will also add it to `LIT::PackageOp`.

This work is part of allowing packages to contain zero or more
targets for which pre-compiled artifacts are included, which is tracked
by #17326. To support that, we'd want to treat pre-compiled artifacts
(the elaborated module bytecode and its standalone archive machine code
bytes) as a pair, so we can have packages include a list or map of
these -- as opposed to two lists, or a list of tuples. It also makes
sense to do this because we sometimes assert that if one of these is
present, the other must be also -- representing that with types here is
a way to enforce this partly at build time.

MODULAR_ORIG_COMMIT_REV_ID: c634ed71fa29a68099ff78939af9bc3f02e27e40