[mojo-lang] Allow references to parameters through structures (#21281)
Teaches the mojo-parser to allow references to parameters through the
structs which they are a field of. Allows `foo[x.static_value]` through
even though `x` is "dynamic" as the expression as a whole is actually a
parameter expression.

To enable this I've loosened the checks of dynamic values when
evaluating the `x.` portion of `x.static_value` this allows the dynamic
value of `x` to be returned. The error is then enforced on the rhs of
the `x.*` expression as the `.*` still has the full check.

---------

Co-authored-by: Mogball <jeff@modular.com>
Co-authored-by: Laszlo Kindrat <laszlo@modular.ai>
MODULAR_ORIG_COMMIT_REV_ID: 3abfe1ea907e8eded6af0c3e7451182731b3cf0b