[GPU] Add ability to load PTX module in Mojo (#22087)
This is another baby step needed to be able to run a CUDA kernel in pure
mojo. To do that, we need to be able to load a CUDA module.

While at it, reorder some of the contents of nvidia_host to make it
easier to navigate the file

GPU_ORIG_REV_ID: 472558e125c50d8bfdb7071a7829d983e16b4d1e