[mojo-lang] Cache the fixed-point iteration of module dependencies (#22174)
Parsing caching requires tracking the dependencies of every module, so
that a change to a dependency triggers a cache miss of the module in
question. This requires resolving the body of the module, looking for
imports.

Then, the parser recurses on the imports to find their dependencies,
and then it does a fixed point iteration to flatten the dependency
graph and uses this to form the cache key for the module.

Only the first step is cached, meaning that even on a cache hit, all
module dependencies are visited and parsed, resulting most of the stdlib
packages getting visited even for an empty file.

Caching the whole process gets closed to a you-get-what-you-pay-for
model, bringing the parse time for an empty Mojo file from 140ms down to
65ms, and reducing the parse time for other files by a smaller amount.

MODULAR_ORIG_COMMIT_REV_ID: 163c17cf259a11b3e7b9d339f4c51e151a7a04c5