[CI] Add Very Basic S3 Intermediate Storage for CI (#22252)
Now that we have some more complicated workflows, we'd like to be able
to take artifacts from one step of the workflow to another step. I would
like to do this without relying on the `upload-artifact` action from
Github for size reasons, so I finally got around to writing a very basic
script that uploads artifacts to S3 for a worklfow.

I am led to believe that the combination of `GITHUB_RUN_ID` and
`GITHUB_RUN_ATTEMPT` is unique across all workflows within a given repo,
so I'm just going to use those three things as a key. This bucket is
versioned, and we can set per-path lifecycle rules if we ever feel like
we need it

MODULAR_ORIG_COMMIT_REV_ID: 7621b95411579ea40103b9f69144043678b92205