Revert "[mojo-lang] Cache the fixed-point iteration of module dependencies (#22174)" (#22245)
This is the wrong thing to do. The transitive dependencies of a file is
not a function of the just the contents of the file, meaning the caching
is incorrect.

As another happy side-effect, this was breaking parser determinism,
because what `resolveModuleDependencies` was doing was visiting all the
dependencies and adding them to the MLIR ModuleOp in one order, but then
on a cache hit, they would get added lazily, resulting in another order.

This reverts commit 163c17cf259a11b3e7b9d339f4c51e151a7a04c5.

MODULAR_ORIG_COMMIT_REV_ID: 3be8b447a684aa82e1c4caf8dc736e757909b2ba