[Support] Add `TargetInfoAttr::relocation_model` (#22218)
Add a relocation model to the `TargetInfoAttr` attribute. This
additional attribute ensures that two targets that are otherwise
identical, but one with relocation model static compared to another with
relocation model PIC, do not compare equal.

This is important when determining whether an existing pre-built
`.mojopkg` can be used for a target. If the package is built for the
same architecture, but a different relocation model, it should not be
used.

Co-authored-by: River Riddle riddleriver@gmail.com
MODULAR_ORIG_COMMIT_REV_ID: 88af9cf45b7a85089b086b46992008b53771461f