[Support] Add LLVMYAMLForwardDecls.h to forward-declare YAMLTraits.h macros (#22236)
`YAMLTraits.h` is a huge header with a big impact on compilation time.
Forward-declaring things from it helps a lot, but some of the macros can
be a little complicated. It's easiest to abstract this out into a header
so we can do this only once.

MODULAR_ORIG_COMMIT_REV_ID: bc6099484c4f020a01c77533f15ed068e3e6602c