[Kernels] Add a lit test runner for GPU kernels (#22283)
This adds a custom lit test runner that will take two files and compile
the gpu kernel for the gpu-target and pass a reference to the cpu file.
This allows end-to-end testing of the GPU code.

There is still some hacks in the testing. For example, we need to pass
`parsing-stdlib` because of #22108 and we need to do some ptx code
cleanup. We also (for some reason) need to allocate a byte on the GPU
and then free it for the kernel to be loaded correctly!!!

All these hacks will be deleted in the next few days

GPU_TEST_ORIG_REV_ID: 6f33402519ca90515736ca841c01b79624f2cce7