[Build] Silence GCC Warnings to make cmake output succinct, NFC (#22326)
One of the benifits of programming for GPUs is that you need an old
compiler for them to work. That compiler also needs to match the how the
Linux driver was built. In the case of Ubuntu, this means using GCC 9.
So, to preserve my sanity make sure that the GCC build has minimal
warnings and things I need to look at.

MODULAR_ORIG_COMMIT_REV_ID: c61ce6850a58e2b29e5225f2f7afb62e502bcbbb