[mojo-driver] Add a debug info language option (#22322)
This option is useful to emit compile unit DIEs with either the C or
Mojo language. The Mojo option is useful for development of the mojo
support in LLDB, and C is useful in the meantime for debuggability.

MODULAR_ORIG_COMMIT_REV_ID: 874e7e7e0e15abd5f437b6ffea8bb7375b7837d6