[KGEN] Make the elaborator and lower closures deterministic (#22311)
In both the elaborator and lower closures, there were issues with
generating unique symbol names and relying on the symbol table to do so,
but because the symbol table is shared across threads, this results in
nondeterministic symbol names.

In addition, the elaborator requires sorting the instantiations of
generators, because parallel traversal of the expansion graph causes the
instantiations to be created in a nondeterministic order.

This ensures `kgen -elaborate` produces deterministic **TEXTUAL** IR,
but there is still an issue with bytecode, caching, and resource names
that has to be fixed.

MODULAR_ORIG_COMMIT_REV_ID: 8867cc72b6a00c62eb89144137c0ce0d8c79ac7b