[mojo-tooling][windows] Fix display of 🔥 in mojo and MojoLLDB help (#22292)
This requires compiling with /utf-8, so that source files are treated as
unicode. We may want to consider enabling this across the codebase,
but we need to validate we don't implicitly rely on the default
codepage (which is tough right now because not everything builds/gets
tested on windows).

See #17121

MODULAR_ORIG_COMMIT_REV_ID: 900f8a60d3e8c793df80bd90b934cab147765696