[mojo-lang] Implement keyword parameter support for direct calls (#22428)
This patch delivers basic support for keyword parameters in functions.
Keyword parameters in indirect calls are not supported yet, and will be
implemented in a subsequent patch. Implementing the binding verification
and overload resolution logic in the presence of keyword parameters
required a richer set of error diagnostics; this is currently in a rough
shape, but at least tested, and will be refactored in subsequent
patches.

MODULAR_ORIG_COMMIT_REV_ID: 50db15cfb19e8e5ede4aee6b37e37285572f85e1