[KGEN][Kernels] Separable compilation via new ParamOperatorAttr (#22551)
This PR is a proof-of-concept of separable compilation by adding a new
parameter operator: `compile_assembly`. It takes two operands, a target
and a signature type, and returns a `!kgen.string`. During elaboration,
a pre-elaboration IR module is sliced out with a similar technique used
for search, the target is overwritten with the selected target, and the
module is run through to assembly via the ObjectCompiler.

MODULAR_ORIG_COMMIT_REV_ID: babda970baf9ff8801a4d15fa469f13fb134994c