[Cache] Add read-only FS cache option (#22602)
Mojo makes great use of a filesystem backend for its cache, but when run
in different environments, we cannot be certain that the Mojo process
can write to the backing cache directory. For example, #21849 caches the
results of Mojo package compilation in the REPL/Jupyter JIT, but the
Mojo playgrounds (Mojohub) run in a locked-down, secure environment that
prevents writing to the filesystem. Using
`Cache::getLocalDefaultBackendChain` in such an environment results in
the filesystem backend being selected for the cache, but any cache write
attempts result in an error.

This patch adds a check whether the filesystem backend is writeable or
readonly. If readonly, filesystem contents are read from, but not
written to. Writes are instead performed to an in-memory backend, which
is read from in addition to the filesystem.

Unit-testing this is a little tricky: a "mock" filesystem would allow us
to simulate a read-only directory, but would involve rewriting a
signifiant portion of the cache logic. Alternatively, I could create a
readonly directory in CMake, but that would require employees to `sudo`
when clearing their build directories. For now, just re-use the existing
cache tests, with plans to stub out the filesystem in a future commit.

MODULAR_ORIG_COMMIT_REV_ID: 4dbfffc58f9a999f8a56551779c105aa4ccd828b